// src/api/definitions.ts
// Generated by https://quicktype.io

export interface SchoolWiresCalendarEvent {
    Id:               number;
    EventID:          number;
    ModuleInstanceID: number;
    RecurringEvent:   boolean;
    RegisteredUsers:  boolean;
    Title:            string;
    Start:            string;
    End:              string;
    CategoryColor:    string;
    CategoryTitle:    string;
    EventSource:      EventSource;
    RecurringInfo:    RecurringInfo;
    AllDay:           boolean;
    NoEndTime:        number;
}

export enum EventSource {
    DistrictCalendar = "District Calendar",
}

export enum RecurringInfo {
    Tbd = "TBD",
}